refactored graph state
