+++ Client/src/Models/Datasets/IDatasetRowModel.tsx
@@ -1,0 +1,15 @@
+export interface IDatasetRowModel {
+  id: number,
+  name: string,
+  isInitiallyHidden: boolean,
+  color: string,
+  shape: string,
+}
+
+export const newDatasetRow: IDatasetRowModel = {
+  id: null,
+  name: null,
+  isInitiallyHidden: false,
+  color: null,
+  shape: null
+}
