+++ Client/src/Models/Graph/IGraphDatasetModel.tsx
@@ -1,4 +1,3 @@
-
 export interface IGraphPoint {
   x: number,
   y: number
@@ -13,6 +12,13 @@ export interface IGraphDatasetModel {
   name: string
 }
 
+export interface IGraphDatasetState {
+  isHidden: boolean,
+  id: number,
+  color: string,
+  shape: string
+}
+
 export const newGraphDataset: IGraphDatasetModel = {
   points: undefined,
   isHidden: false,
@@ -20,4 +26,13 @@ export const newGraphDataset: IGraphDatasetModel = {
   color: null,
   shape: null,
   name: null
+}
+
+export const toGraphDatasetState = (dataset: IGraphDatasetModel): IGraphDatasetState => {
+  return {
+    color: dataset.color,
+    id: dataset.id,
+    isHidden: dataset.isHidden,
+    shape: dataset.shape
+  }
 }
