+++ Client/src/Models/Graph/IGraphStateModel.tsx
@@ -1,15 +1,10 @@
+import { IGraphDatasetState } from "./IGraphDatasetModel"
+
 export interface IGraphStateModel {
-  datasets: IDisplayedDatasetStateModel[],
+  datasets: IGraphDatasetState[],
   axes: IAxisStateModel[],
   name: string,
   id: string
-}
-
-export interface IDisplayedDatasetStateModel {
-  id: number,
-  isHidden: boolean,
-  color: string,
-  shape: string,
 }
 
 export interface IAxisStateModel {
@@ -20,28 +15,12 @@ export interface IAxisStateModel {
   units: string
 }
 
-export interface IDatasetRowModel {
-  id: number,
-  name: string,
-  isInitiallyHidden: boolean,
-  color: string,
-  shape: string,
-}
-
 export const newAxisState: IAxisStateModel = {
   logarithmic: false,
   units: 'mm',
   variableName: 'cell width',
   zoomStartIndex: null,
   zoomEndIndex: null,
-}
-
-export const newDatasetRow: IDatasetRowModel = {
-  id: null,
-  name: null,
-  isInitiallyHidden: false,
-  color: null,
-  shape: null
 }
 
 export const newGraphState: IGraphStateModel = {
